refactor(test): remove all session api related integration-tests
remove all session api related integration tests
