+++ packages/integration-tests/src/api/index.ts
@@ -3,7 +3,6 @@ export * from './connector.js';
 export * from './application.js';
 export * from './sign-in-experience.js';
 export * from './admin-user.js';
-export * from './session.js';
 export * from './logs.js';
 export * from './dashboard.js';
 export * from './me.js';
