+++ packages/integration-tests/src/tests/api/interaction/sign-in-with-passcode-identifier.test.ts
@@ -9,13 +9,16 @@ import {
   deleteUser,
   updateSignInExperience,
 } from '#src/api/index.js';
-import { expectRejects, readPasscode } from '#src/helpers.js';
+import { initClient, processSession, logoutClient } from '#src/helpers/client.js';
+import {
+  clearConnectorsByTypes,
+  setEmailConnector,
+  setSmsConnector,
+} from '#src/helpers/connector.js';
+import { expectRejects, readPasscode } from '#src/helpers/index.js';
+import { enableAllVerificationCodeSignInMethods } from '#src/helpers/sign-in-experience.js';
+import { generateNewUser, generateNewUserProfile } from '#src/helpers/user.js';
 import { generateEmail, generatePhone } from '#src/utils.js';
-
-import { initClient, processSession, logoutClient } from './utils/client.js';
-import { clearConnectorsByTypes, setEmailConnector, setSmsConnector } from './utils/connector.js';
-import { enableAllVerificationCodeSignInMethods } from './utils/sign-in-experience.js';
-import { generateNewUser, generateNewUserProfile } from './utils/user.js';
 
 describe('Sign-In flow using verification-code identifiers', () => {
   beforeAll(async () => {
